feat: :money_with_wings: create a wallet module

+++ src/core/users/entities/user.entity.ts
@@ -51,8 +51,8 @@ export class UserEntity extends BaseEntity {
   @Property({ type: 'integer', default: 0 })
   left_count = 0;
 
-  @Property({ type: 'integer', default: 0 })
-  coins = 0;
+  @Property({ type: 'bigint', default: 0 })
+  coins = 0n;
 
   @Property({ type: 'timestamptz', nullable: true })
   birth_date: Date | null = null;
